Seperation of concern
